简单的游玩界面实现

+++ Assets/Scripts/Core/Tile.cs
@@ -2,14 +2,15 @@ namespace WaveTeam.Core
 {
     /// <summary>
     /// 波形的一段（一个 step）。三个属性：音高(高度)、时值(拍数)、特质(形状/音色)。
+    /// 时值为浮点拍数：0.5=八分，1=四分，1.5=附点四分，2=二分。
     /// </summary>
     public sealed class Tile
     {
-        public readonly int Pitch;    // 音高，越高 tile 越高
-        public readonly int Duration; // 时值（拍数），1 = 一拍，2 = 两拍
-        public readonly Trait Trait;  // 特质，决定形状与音色
+        public readonly int Pitch;      // 音高，越高 tile 越高
+        public readonly float Duration; // 时值（拍数）：0.5=八分, 1=四分, 1.5=附点四分, 2=二分
+        public readonly Trait Trait;    // 特质，决定形状与音色
 
-        public Tile(int pitch, int duration, Trait trait)
+        public Tile(int pitch, float duration, Trait trait)
         {
             Pitch = pitch;
             Duration = duration;
